+++ src/hooks/review/useAddComment.js
@@ -1,33 +1,22 @@
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { collection, serverTimestamp, addDoc } from "firebase/firestore";
 import { useSelector } from "react-redux";
-import { db } from "@/services/firebase";
 import { QUERY_KEYS } from "@/constants/queryKeys";
+import { useMutationWithFeedback } from "@/hooks/common/useMutationWithFeedback";
+import { addComment } from "@/services/review/addComment";
 
 export const useAddComment = (reviewId) => {
-  const queryClient = useQueryClient();
   const user = useSelector((state) => state.user.user);
 
-  return useMutation({
-    mutationFn: async ({ content }) => {
-      if (!user) throw new Error("로그인이 필요합니다.");
+  const checkAuth = () => {
+    if (!user) throw new Error("로그인이 필요합니다.");
+  };
 
-      const commentRef = collection(db, "reviews", reviewId, "comments");
-      await addDoc(commentRef, {
-        content,
-        author: {
-          uid: user.uid,
-          displayName: user.displayName || "익명",
-          photoURL: user.photoURL || "",
-        },
-        authorUid: user.uid,
-        createdAt: serverTimestamp(),
-      });
+  return useMutationWithFeedback({
+    mutationFn: async ({ content }) => {
+      checkAuth();
+      return await addComment({ reviewId, content, user });
     },
-    onSuccess: () => {
-      queryClient.invalidateQueries({
-        queryKey: QUERY_KEYS.COMMENTS(reviewId),
-      });
-    },
+    successMessage: "댓글이 등록되었습니다.",
+    errorMessage: "댓글 등록 중 오류가 발생했습니다.",
+    queryKeysToInvalidate: [[QUERY_KEYS.COMMENTS(reviewId)]],
   });
 };
