+++ src/hooks/review/useDeleteComment.js
@@ -1,17 +1,13 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { doc, deleteDoc } from "firebase/firestore";
-import { db } from "@/services/firebase";
 import { QUERY_KEYS } from "@/constants/queryKeys";
+import { deleteComment } from "@/services/review/deleteComment";
 
 export const useDeleteComment = (reviewId) => {
   const queryClient = useQueryClient();
 
   return useMutation({
-    mutationFn: async (commentId) => {
-      const commentDoc = doc(db, "reviews", reviewId, "comments", commentId);
-      await deleteDoc(commentDoc);
-    },
-    // Optimistic Update
+    mutationFn: (commentId) => deleteComment({ reviewId, commentId }),
+
     onMutate: async (commentId) => {
       await queryClient.cancelQueries({
         queryKey: QUERY_KEYS.COMMENTS(reviewId),
@@ -28,7 +24,6 @@ export const useDeleteComment = (reviewId) => {
       return { prevComments };
     },
 
-    // 실패 시 롤백
     onError: (err, commentId, context) => {
       if (context?.prevComments) {
         queryClient.setQueryData(
@@ -38,7 +33,6 @@ export const useDeleteComment = (reviewId) => {
       }
     },
 
-    // 성공 시 re-fetch로 동기화
     onSettled: () => {
       queryClient.invalidateQueries({
         queryKey: QUERY_KEYS.COMMENTS(reviewId),
